fix: move configs into App
