+++ packages/legacy/core/App/configs/ledgers/indy/__mocks__/fixtures/sovrin-main-net-pool.json
