+++ packages/legacy/core/App/defaultConfiguration.ts
@@ -2,7 +2,7 @@ import { useProofRequestTemplates } from '@hyperledger/aries-bifold-verifier'
 import { IOverlayBundleData } from '@hyperledger/aries-oca'
 import { BrandingOverlayType, DefaultOCABundleResolver } from '@hyperledger/aries-oca/build/legacy'
 
-import defaultIndyLedgers from '../configs/ledgers/indy'
+import defaultIndyLedgers from './configs/ledgers/indy'
 
 import * as bundle from './assets/oca-bundles.json'
 import EmptyList from './components/misc/EmptyList'
