+++ packages/legacy/core/App/utils/ledger.tsx
@@ -1,6 +1,6 @@
 import TcpSocket from 'react-native-tcp-socket'
 
-import pools from '../../configs/ledgers/indy'
+import pools from '../configs/ledgers/indy'
 import { GenesisTransaction } from '../types/genesis'
 
 export const canConnectToLedgerNode = async (node: { host: string; port: number }): Promise<boolean> =>
